third_party/WebKit/LayoutTests/external/wpt/webvtt/parsing/file-parsing/tests/regions-old-expected.txt

This is a testharness.js-based test.
FAIL regions, old assert_equals: expected null but got object "[object VTTRegion]"
Harness: the test ran to completion.
